+++ src/main/java/com/loja/loja/DTO/ProdutoDTO.java
@@ -1,5 +1,7 @@
 package com.loja.loja.DTO;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.Setter;
 
@@ -10,7 +12,7 @@ public class ProdutoDTO {
     private Integer id;
     private String nome;
     private Integer quantidade;
-    private Integer valor;
+    private BigDecimal valor;
     private Integer idVendedor;
     private VendedorProdutoDTO vendedorProduto;
